API reference

Index

AtomsBase.AbstractSystem

AtomsBase.atomic_mass

AtomsBase.atomic_number

AtomsBase.atomic_symbol

AtomsBase.boundary_conditions

AtomsBase.bounding_box

AtomsBase.chemical_formula

AtomsBase.element

AtomsBase.n_dimensions

AtomsBase.periodicity

AtomsBase.species_type

AtomsBase.velocity

Base.position

Types

AtomsBase.AbstractSystem — Type

AbstractSystem{D}

A D-dimensional system.

Functions

AtomsBase.atomic_mass — Function

atomic_mass(sys::AbstractSystem)
atomic_mass(sys::AbstractSystem, i)
atomic_mass(species)

Vector of atomic masses in the system sys or the atomic mass of a particular species / the ith species in sys. The elements are <: Unitful.Mass.

AtomsBase.atomic_number — Function

atomic_number(sys::AbstractSystem)
atomic_number(sys::AbstractSystem, i)
atomic_number(species)

Vector of atomic numbers in the system sys or the atomic number of a particular species / the ith species in sys.

AtomsBase.atomic_symbol — Function

atomic_symbol(sys::AbstractSystem)
atomic_symbol(sys::AbstractSystem, i)
atomic_symbol(species)

Vector of atomic symbols in the system sys or the atomic symbol of a particular species / the ith species in sys.

AtomsBase.chemical_formula — Function

Returns the chemical formula of an AbstractSystem as a string.

AtomsBase.boundary_conditions — Function

boundary_conditions(sys::AbstractSystem{D})

Return a vector of length D of BoundaryCondition objects, one for each direction described by bounding_box(sys).

AtomsBase.bounding_box — Function

bounding_box(sys::AbstractSystem{D})

Return a vector of length D of vectors of length D that describe the "box" in which the system sys is defined.

AtomsBase.element — Function

The element corresponding to a species/atom (or missing).

AtomsBase.n_dimensions — Function

n_dimensions(::AbstractSystem)
n_dimensions(atom)

Return number of dimensions.

AtomsBase.periodicity — Function

Return vector indicating whether the system is periodic along a dimension.

AtomsBase.species_type — Function

species_type(::AbstractSystem)

Return the type used to represent a species or atom.

AtomsBase.velocity — Function

velocity(sys::AbstractSystem{D})
velocity(sys::AbstractSystem, index)
velocity(species)

Return a vector of velocities of every particle in the system sys. Return type should be a vector of vectors each containing D elements that are <:Unitful.Velocity. If an index is passed or the action is on a species, return only the velocity of the referenced species. Returned value of the function may be missing.

Base.position — Function

position(sys::AbstractSystem{D})
position(sys::AbstractSystem, index)
position(species)

Return a vector of positions of every particle in the system sys. Return type should be a vector of vectors each containing D elements that are <:Unitful.Length. If an index is passed or the action is on a species, return only the position of the referenced species / species on that index.